Remove CertificatePinningRegistry and merge ServerTrustRegistry into CertificatePinning.swift as a breaking change
